feat: implement logging of method execution time with @LogExecutionTime annotation and aspect #126 refactor: update ride matching logic to use new location fields and improve detour calculations fix: enhance error response with detailed message and update global exception handling

+++ code/backend/src/main/java/com/br/puc/carona/service/RouteDetails.java
@@ -3,4 +3,4 @@ package com.br.puc.carona.service;
 /**
  * Record holding route calculation details.
  */
-public record RouteDetails(double totalDistance, int totalSeconds) {}
+public record RouteDetails(double totalDistance, double totalSeconds) {}
